// src/actions/create-appointment/schema.ts
import { z } from "zod";

export const createAppointmentSchema = z.object({
  patientId: z.string().uuid({ message: "Paciente é obrigatório" }),
  doctorId: z.string().uuid({ message: "Médico é obrigatório" }),
  date: z.date({ message: "Data é obrigatória" }),
  time: z.string().min(1, { message: "Horário é obrigatório" }),
  appointmentPriceInCents: z
    .number()
    .min(1, { message: "Preço da consulta é obrigatório" }),
});

export type CreateAppointmentSchema = z.infer<typeof createAppointmentSchema>;